+++ src/renderer/components/Viewpoints/ViewpointSelection.tsx
@@ -148,6 +148,10 @@ export default function ViewpointSelection(props: IProps) {
           multiPlayerMode: s.length > 1 ? mode : false,
           // Always pause if changing selections in multiplayer mode.
           playing: multiPlayerMode ? false : playing,
+          // Update the preferred viewpoint to the newly selected player. This
+          // means  the video table will also update so switching between videos
+          // will stick to the same player.
+          preferredViewpoint: selection.player?._name || '',
         };
       });
     };